chore: - Change all Repository method IDs to Long   (NotificationRepository, StatusRepository, TaskAssigneeRepository,    TaskHistoryRepository, TaskRepository, UserRepository, WorkflowRepository) - Rename UserRepository method: findByUserId -> findById - Update WorkflowMapper.dtoToEntity: findByUserId -> findById
test:
- Create WorkflowServiceImplTest

note:
- Many errors occurred before fixing; do these changes first
- Tests not yet implemented

+++ TaskFlowManagementSystem-backend/src/main/java/com/example/demo/repository/NotificationRepository.java
@@ -12,10 +12,10 @@ import io.lettuce.core.dynamic.annotation.Param;
 
 public interface NotificationRepository extends JpaRepository<Notification, Long>{
 	 // 查某任務所有被指派者
-    List<TaskAssignee> findByTaskId(Integer taskId);
+    List<TaskAssignee> findByTaskId(Long taskId);
 
     // 查某使用者被指派的任務
-    List<TaskAssignee> findByUserId(Integer userId);
+    List<TaskAssignee> findByUserId(Long userId);
 
     // 查某使用者負責的某個任務節點
     @Query(value = "SELECT n.id AS notification_id, n.message, n.read_flag, n.created_at, " +
